refactor: kakomap api 모듈 분리(domain -> infra)
