+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/crag/domain/ExternalMapAdapter.kt
@@ -1,0 +1,7 @@
+package org.depromeet.clog.server.domain.crag.domain
+
+import org.depromeet.clog.server.domain.crag.dto.KakaoSearchResponseDto
+
+interface ExternalMapAdapter {
+    fun searchKakaoLocationResult(query: String, page: Int): KakaoSearchResponseDto?
+}
